requirements.txt
albumentations==1.3.0
bunch==1.0.1
loguru==0.7.0
matplotlib==3.7.1
numpy==1.24.3
einops==0.6.1
opencv-python-headless==4.7.0.72
Pillow==9.5.0
ruamel.base==1.0.0
ruamel.yaml==0.17.32
scikit-learn==1.2.2
tensorboard==2.13.0
timm==0.9.2
torch==2.0.1
torchstat==0.0.7
torchvision==0.15.2
tqdm==4.61.1
ttach==0.0.3
yacs==0.1.8
chardet==5.1.0
efficientnet-pytorch==0.7.1
imageio==2.31.0
ipykernel==6.23.1
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.23.0
kiwisolver==1.4.4
matplotlib-inline==0.1.6
segmentation-models-pytorch==0.3.3
